Refactor/mtf (#197)
* Refactor MTF module with BaseMTF and shared logic

This commit completes the refactoring of the MTF functionalities.
Key changes include:

1.  **Renamed `MTFBase` to `BaseMTF`**.
2.  **Enhanced `BaseMTF` (`optiland/mtf/base.py`)**:
    *   Added abstract methods: `_generate_mtf_data`, `_calculate_psf`,
        and `_plot_field_mtf`.
    *   Implemented a concrete `view` method that handles common plotting
        setup, including the diffraction limit reference.
    *   Moved logic for resolving "all" fields and "primary" wavelength
        into `BaseMTF.__init__`.
3.  **Refactored `FFTMTF` (`optiland/mtf/fft.py`)**:
    *   Inherits from the enhanced `BaseMTF`.
    *   Uses `BaseMTF` for field/wavelength resolution.
    *   PSF calculation moved to `_calculate_psf` method.
    *   Implements `_plot_field_mtf` for its specific plotting needs,
        relying on `BaseMTF.view()` for the overall plot structure.
    *   `self.freq` is now calculated in `__init__`.
4.  **`GeometricMTF` and `SampledMTF`**: Remain as previously refactored,
    without inheriting from `BaseMTF`, as per design decisions.
5.  **Tests (`tests/test_mtf.py`)**: Verified to be compatible with
    the changes to `FFTMTF`'s view mechanism.

Note: I repeatedly tried and failed to run `ruff` for formatting and linting checks due to persistent environment issues. The code changes are submitted with the understanding that minor linting/formatting adjustments might be needed if `ruff` identifies issues in a different environment.

* chore: ruff formatting

* fix: update import path for SampledMTF in through_focus_mtf.py

* refactor: update docstrings in base.py and fft.py for clarity and consistency

* refactor: simplify wavelength handling in SampledMTF initialization

---------
